refactor: move coverage intelligence contracts to output crate
